Impove the logic of buttons and the mouse panel.
Fix some bugs.

Now lakes will generated on the path map.

Now the signal power calculation for users are replaced with a more complicated algorithm when the path map is open, so the FPS may be extremely low if lots of them are connected to the net service and the analysis mode is on. But this change is necessary for future updates, where the decay constant will various on the path map.